+++ src/Animations/index.ts
@@ -1,0 +1,1 @@
+export {default as LaptopAnimation} from "./laptop-animation.json";