+++ src/Components/index.ts
@@ -1,0 +1,2 @@
+export {default as NavLinks} from "./NavLinks/NavLinks";
+export {default as AboutMe} from "./AboutMe/AboutMe";